fix: Fix arrays, add new opcodes STORE, ASTORE, ALOAD
An issue that we encountered was that we had the base address and we knew how much we needed to offset from that address to access the value from an array, but we didn't have a way to distinguish an address reference from a global variable.

A flag was added called ADDRESS_REFERENCE that lives in the 26th bit of the address. When it's turned to 1, it will indicate that the address references another address and not a direct variable.

This could possibly be refactored and improved but this was the initial approach and it works well for now.

When we want to pop an operand from the stack and it's an address reference, if we want to do an assignment, it will insert the assignment with the opcode ASTORE. If it's not an assignment and we want to access the variable, it will insert an ALOAD operation and return that address.